make.sh/clean.sh: Misc updates

+++ clean.sh
@@ -1,3 +1,9 @@
 #!/bin/sh
-rm -rf build/
-mkdir build
+BUILD="build/"
+RELEASE="release/"
+if [ -d "$BUILD" ]; then
+  rm -rf build/
+fi
+if [ -d "$RELEASE" ]; then
+  rm -rf release/
+fi
